@@ -1,18 +1,42 @@
 module github.com/GoogleCloudPlatform/berglas
 
-go 1.14
+go 1.17
 
 require (
-	cloud.google.com/go v0.60.0
-	cloud.google.com/go/storage v1.10.0
-	github.com/gammazero/workerpool v0.0.0-20200608033439-1a5ca90a5753
-	github.com/gofrs/uuid v3.3.0+incompatible
-	github.com/golang/protobuf v1.4.2
-	github.com/pkg/errors v0.9.1
-	github.com/sethvargo/go-retry v0.1.0
-	github.com/sirupsen/logrus v1.6.0
-	github.com/spf13/cobra v1.0.0
-	google.golang.org/api v0.29.0
-	google.golang.org/genproto v0.0.0-20200711021454-869866162049
-	google.golang.org/grpc v1.30.0
+	cloud.google.com/go/iam v0.1.1
+	cloud.google.com/go/kms v1.2.0
+	cloud.google.com/go/secretmanager v1.1.0
+	cloud.google.com/go/storage v1.20.0
+	github.com/client9/misspell v0.3.4
+	github.com/gammazero/workerpool v1.1.2
+	github.com/sethvargo/go-retry v0.2.2
+	github.com/sirupsen/logrus v1.8.1
+	github.com/spf13/cobra v1.3.0
+	golang.org/x/tools v0.1.5
+	google.golang.org/api v0.67.0
+	google.golang.org/genproto v0.0.0-20220207185906-7721543eae58
+	google.golang.org/grpc v1.44.0
+	google.golang.org/protobuf v1.27.1
+	honnef.co/go/tools v0.0.1-2020.1.4
+)
+
+require (
+	cloud.google.com/go v0.100.2 // indirect
+	cloud.google.com/go/compute v1.2.0 // indirect
+	github.com/BurntSushi/toml v0.3.1 // indirect
+	github.com/gammazero/deque v0.1.0 // indirect
+	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
+	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/google/go-cmp v0.5.7 // indirect
+	github.com/googleapis/gax-go/v2 v2.1.1 // indirect
+	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	go.opencensus.io v0.23.0 // indirect
+	golang.org/x/mod v0.5.0 // indirect
+	golang.org/x/net v0.0.0-20220127200216-cd36cc0744dd // indirect
+	golang.org/x/oauth2 v0.0.0-20211104180415-d3ed0bb246c8 // indirect
+	golang.org/x/sys v0.0.0-20220204135822-1c1b9b1eba6a // indirect
+	golang.org/x/text v0.3.7 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	google.golang.org/appengine v1.6.7 // indirect
 )
