@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// +build tools
+//go:build tools
 
 // Package tools includes the list of tools used in the project.
 package tools
